Major refactoring to use subclasses
