+++ cappa/exceptions.py
@@ -1,0 +1,8 @@
+class MissingExecutable(Exception):
+    """Exception raised when a package manager executable is missing"""
+    pass
+
+
+class UnknownManager(Exception):
+    """Exception raised when a requested package manager is unknown"""
+    pass
