+++ setup.py
@@ -1,13 +1,13 @@
-from setuptools import setup
+from setuptools import setup, find_packages
 
 setup(
     name="cappa",
-    version="0.13",
+    version="0.14",
     description="Package installer for Captricity. Supports apt-get, pip, bower, and npm.",
     author="Yoriyasu Yano",
     author_email="yorinasub17@gmail.com",
 
-    py_modules=["cappa"],
+    packages=find_packages(),
     scripts=["scripts/cappa"],
     install_requires=open("requirements.txt").read().split(),
     tests_requires=open("test_requirements.txt").read().split(),
